requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
google-genai>=1.57.0
python-dotenv>=1.2.1
python-multipart>=0.0.21
numpy>=2.4.0
pillow>=12.1.0
rasterio>=1.5.0
fal-client>=0.11.0
requests>=2.31.0
